Base MemoryMappedFile implementation on common OSFile interface. Add test case for ARROW-340.
