+++ cpp/src/arrow/io/test-common.h
@@ -33,6 +33,7 @@
 #endif
 
 #include "arrow/buffer.h"
+#include "arrow/io/file.h"
 #include "arrow/io/memory.h"
 #include "arrow/memory_pool.h"
 #include "arrow/test-util.h"
